+++ app/controllers/admin/admin_products_controller.ts
@@ -1,6 +1,8 @@
 import type { HttpContext } from '@adonisjs/core/http'
+import app from '@adonisjs/core/services/app'
 import Product from "#models/product";
 import { AdminCreateProductValidator, AdminUpdateProductValidator } from '#validators/admin_product_validator';
+import { cuid } from '@adonisjs/core/helpers';
 
 export default class AdminProductsController{
     async index({view}:HttpContext){
@@ -15,8 +17,12 @@ export default class AdminProductsController{
 
     async store({request,response}:HttpContext){
         const data = await request.validateUsing(AdminCreateProductValidator)
-        // const data = request.only(['name', 'description', 'price', 'stock', 'imageUrl'])
-        await Product.create(data)
+        const {image, ...tempProduct} = data
+        if (image){
+            await image.move(app.makePath('storage/uploads'),{name:`${cuid()}.${image.extname}`});
+            (tempProduct as any).imageUrl = `/storage/uploads/${image.fileName}`
+        }
+        await Product.create(tempProduct)
         return response.redirect('/admin/products')
     }
 
@@ -27,9 +33,14 @@ export default class AdminProductsController{
 
     async update({params, request, response}:HttpContext){
         const data = await request.validateUsing(AdminUpdateProductValidator)
+        const {image, ...tempProduct} = data
         const product=await Product.findOrFail(params.id)
-        // const data = request.only(['name', 'description', 'price', 'stock', 'imageUrl'])
-        product.merge(data)
+        if (image){
+            const fileName = `${cuid()}.${image.extname}`
+            await image.move(app.makePath(`storage/uploads`), {name:fileName}) 
+            product.imageUrl=`/storage/uploads/${fileName}`
+        }
+        product.merge(tempProduct)
         await product.save()
         return response.redirect('/admin/products')
     }